requirements.txt
pandas
tables
numpy
tqdm
einops
torchcde
scikit-learn
networkx
gensim
node2vec
torch-scatter
pytorch-msssim